+++ R/drawCell.R
@@ -1,48 +1,42 @@
 #' Creates a cell picture
 #'
-#' @param organism_identifier Taxonomy id of the organism of interest.
-#'   Can be found at https://www.ncbi.nlm.nih.gov/taxonomy/
-#' @param sl_ids Id of the subcellular location, can be found at
-#'  http://current.geneontology.org/ontology/external2go/uniprotkb_sl2go
-#' @param inputId input id of cell' defaults to `NULL`
-#' @param width width of cell; defaults to `NULL`
-#' @param height height of cell; defaults to `NULL`
-#' @param elementId id of cell, defaults to `NULL`
-#' @param colour_sl Highlighted color of the subcelullar components
+#' @param organism_identifier Taxonomy id of the organism of interest. Can be found at
+#' https://www.ncbi.nlm.nih.gov/taxonomy/
+#' @param list_sl_colors Named list. The name is the id of the subcellular location, can be found at
+#'  http://current.geneontology.org/ontology/external2go/uniprotkb_sl2go the value is the colour that
+#'  each subcellular location should have
+#'  @param width  Width of the resulting image
+#'  @param height Heigth of the resulting image
 #'
 #' @import htmlwidgets
 #'
-#' @return Generates a PNG of the requested cell by the user with highlighted
+#' @return Generates a vector image of the requested cell by the user with coloured
 #' subcellular elements.
 #' @export
 #'
 #' @examples
-#' drawCell(organism_identifier = '58334', sl_ids = '0049', color = 'lightblue')
-#'
-drawCell <-
-  function(
-    organism_identifier,
-    sl_ids,
-    inputId = NULL,
-    width = NULL,
-    height = NULL,
-    elementId = NULL,
-    colour_sl = NULL
-  ) {
-    # forward options using x
-    x <- list(
-      input_id = inputId,
-      organism_identifier = organism_identifier,
-      sl_ids = csl(sl_ids),
-      colour_sl = csl(colour_sl)
-    )
-    # create widget
-    htmlwidgets::createWidget(
-      name = "drawCell",
-      x,
-      width = width,
-      height = height,
-      package = "drawCell",
-      elementId = elementId
-    )
-  }
+#' drawCell(
+#'    organism_identifier = '9606',
+#'    list_sl_colors = list(
+#'    'SL0095' = 'green',
+#'    'SL0154'='red')
+#'  )
+drawCell <- function(organism_identifier, list_sl_colors,  width = NULL, height = NULL) {
+
+  # forward options using x
+  x = list(
+    input_id = "cell",
+    organism_identifier = organism_identifier,
+    sl_ids = drawCell:::csl(names(list_sl_colors)),
+    colour_sl = drawCell:::csl(list_sl_colors)
+  )
+  # create widget
+  htmlwidgets::createWidget(
+    name = 'drawCell',
+    x,
+    width = width,
+    height = height,
+    package = 'drawCell',
+    elementId = NULL
+  )
+}