+++ R/drawCellShiny.R
@@ -1,11 +1,12 @@
 #' Run Shiny app for the drawCell Package
 #'
-#' @return
 #' @export
 #'
-#' @examples
-drawCellShiny <- function() {
-  shiny::shinyAppDir(appDir = base::system.file("shinyApp/drawCellShiny",
-    package = "drawCell"
-  ))
+#' @import shiny
+#' @import shiny.semantic
+#' @import htmltools
+#' @import glue
+drawCellShiny <- function(){
+    shiny::shinyAppDir(appDir = base::system.file('shinyApp/drawCellShiny',
+                                                  package = 'drawCell'))
 }