+++ inst/shinyApp/drawCellShiny/ui.R
@@ -1,14 +1,14 @@
-dashboardPage(
-  dashboardHeader(
-    title = titlePanel(
-      title = h2("drawCell"),
-      windowTitle = tags$head(
-        tags$link(
-          rel = "icon",
-          type = "image/png",
-          href = "favicon.ico"
-        ),
-        tags$title("drawCell")
+semanticPage(
+  includeCSS("www/styles.css"),
+  # Application title
+  titlePanel(
+    title = div(
+      class = "ui massive fluid teal right icon label",
+      "drawCell",
+      a(
+        icon(class = "big github square"),
+        href = "https://github.com/svalvaro/drawCell",
+        target = "_blank"
       )
     ),
     tags$li(
@@ -20,31 +20,82 @@ dashboardPage(
       class = "dropdown"
     )
   ),
-  dashboardSidebar(
-    width = 300,
-    textInput(
-      inputId = "taxIdInput",
-      label = h3("Introduce the species name or Taxonomy ID"),
-      value = "",
-      placeholder = "Homo sapiens / 6096"
+  div(
+    class = "ui attached segments",
+    segment(
+      class = "padded",
+      div(
+        class = "ui horizontal stackable segments",
+        segment(
+          class = "compact padded",
+          text_input(
+            input_id = "taxIdInput",
+            label = label(
+              class = "teal pointing below",
+              "Introduce the species name or Taxonomy ID (Leave blank when selecting cell type)"
+            ),
+            value = "",
+            placeholder = "Homo sapiens / 6096"
+          )
+        ),
+        segment(
+          class = "compact",
+          div(class = "ui vertical divider", "OR")
+        ),
+        segment(
+          class = "compact padded",
+          selectInput(
+            inputId = "cell_type",
+            label = label(
+              class = "teal pointing below",
+              "Select Cell Type (Leave blank when using species name or taxonomy ID)"
+            ),
+            choices = c("", sort(unique(data$Image.name))),
+            selected = "Animal cell",
+            type = "big search clearable selection"
+          )
+        )
+      )
     ),
-    colourpicker::colourInput(
-      inputId = "colourInput",
-      h3("Select colour:"),
-      "#56B4E9",
-      palette = "square",
-      returnName = TRUE,
-      showColour = c("background")
-    ),
-    selectInput("cell_type",
-      h3("Type of cell:"),
-      choices = c("", unique(data$Image.name))
-    )
-  ),
-  dashboardBody(
-    drawCell::drawCellOutput("cell_output"),
-    wellPanel(
-      textOutput("cell_s_output")
+    segment(
+      class = "raised very padded container",
+      label(textOutput("selected_cell_name"), class = "big teal top attached"),
+      div(
+        class = "ui horizontal stackable segments",
+        segment(
+          class = "padded compact",
+          div(
+            class = "ui vertical compact segments",
+            segment(
+              class = "compact padded",
+              colourpicker::colourInput(
+                inputId = "colourInput",
+                label = label(
+                  class = "teal pointing below",
+                  "Select subcellular location to highlight then select color"
+                ),
+                "#56B4E9",
+                palette = "square",
+                returnName = TRUE,
+                showColour = c("background")
+              ),
+              button(
+                input_id = "clear_color",
+                class = "basic",
+                label = "Clear Colors"
+              )
+            ),
+            segment(
+              class = "compact padded",
+              semantic_DTOutput("cell_sl_color")
+            )
+          )
+        ),
+        segment(
+          class = "padded",
+          drawCell:::drawCellOutput("cell_output", height = "100%", width = "100%")
+        )
+      )
     )
   )
 )