+++ NEWS.md
@@ -2,3 +2,4 @@
 
 * Returns cell as interactive `svg`.
 * Submission to Bioconductor 3.14
+* Improve shiny app with `shiny.semantic`